+++ src/vmem/vmem.c
@@ -17,7 +17,7 @@ extern int __start_vmem, __stop_vmem;
     We therefore use __attribute__((weak)) so we can compile in the absence of these. */
 extern __attribute__((weak)) int __start_vmem, __stop_vmem;
 
-void * vmem_memcpy(void * to, void * from, size_t size) {
+void * vmem_memcpy(void * to, void * from, uint32_t size) {
 
 	for(vmem_t * vmem = (vmem_t *) &__start_vmem; vmem < (vmem_t *) &__stop_vmem; vmem++) {
