+++ src/vmem/vmem_fram_cache.c
@@ -14,11 +14,11 @@
  * Driver API declaration from:
  * <drivers/fram.h>
  */
-void fram_write_data(uint16_t addr, void *data, uint16_t len);
-void fram_read_data(uint16_t addr, void *data, uint16_t len);
+void fram_write_data(uint32_t addr, void *data, uint32_t len);
+void fram_read_data(uint32_t addr, void *data, uint32_t len);
 
 
-void vmem_fram_cache_read(vmem_t * vmem, uint32_t addr, void * dataout, int len) {
+void vmem_fram_cache_read(vmem_t * vmem, uint32_t addr, void * dataout, uint32_t len) {
 
 	vmem_fram_cache_driver_t * driver = vmem->driver;
 
@@ -34,7 +34,7 @@ void vmem_fram_cache_read(vmem_t * vmem, uint32_t addr, void * dataout, int len)
 
 }
 
-void vmem_fram_cache_write(vmem_t * vmem, uint32_t addr, void * datain, int len) {
+void vmem_fram_cache_write(vmem_t * vmem, uint32_t addr, void * datain, uint32_t len) {
 
 	vmem_fram_cache_driver_t * driver = vmem->driver;
